feat: add support for snapshot calculations
Supported events are for entitlements and ingestion events. This is
a naive implementation in the sense that we are not caching anything.

We add a basic highwatermark cache to prevent recalculations multiple
times. This will be more important when the ingest part is ready.
